ids_hybride_project/requirements.txt
# IDS Hybride - Dépendances Python

# Web Framework
Flask==2.3.3
Werkzeug==2.3.7

# Machine Learning
scikit-learn==1.3.0
numpy==1.24.3
pandas==2.0.3
xgboost==1.7.6

# Network Capture
pyshark==0.6
scapy==2.5.0

# Database
mysql-connector-python==8.1.0

# Serialization
joblib==1.3.1

# Utilities
python-dotenv==1.0.0
